@@ -1,3 +1,8 @@
+<script setup>
+import Validator from './components/Validator.vue'
+import Footer from './components/Footer.vue'
+</script>
+
 <template>
   <div>
     <b-container>
@@ -5,7 +10,7 @@
         <b-col>
           <b-row class="mb-2">
             <b-col class="flex-grow-0">
-              <img src="/gbfs_validator.jpg" alt="">
+              <img src="/gbfs_validator.jpg" alt="" />
             </b-col>
             <b-col>
               <h1>gbfs-validator.netlify.com</h1>
@@ -14,39 +19,34 @@
           </b-row>
           <b-row>
             <b-col>
-              <p>Validate any <a href="https://github.com/NABSA/gbfs/">GBFS</a> feed online. Based on official <a href="https://github.com/MobilityData/gbfs-json-schema">GBFS json schemas</a>. <br>
-          More details on <a href="https://github.com/MobilityData/gbfs-validator">https://github.com/MobilityData/gbfs-validator</a>.</p>
+              <p>
+                Validate any
+                <a href="https://github.com/NABSA/gbfs/">GBFS</a> feed online.
+                Based on official
+                <a href="https://github.com/MobilityData/gbfs-json-schema"
+                  >GBFS json schemas</a
+                >. <br />
+                More details on
+                <a href="https://github.com/MobilityData/gbfs-validator"
+                  >https://github.com/MobilityData/gbfs-validator</a
+                >.
+              </p>
             </b-col>
           </b-row>
         </b-col>
       </b-row>
       <b-row>
         <b-col>
-          <Validator/>
+          <Validator />
         </b-col>
       </b-row>
       <b-row>
-        <b-col>
-
-        </b-col>
+        <b-col> </b-col>
       </b-row>
     </b-container>
-    <Footer/>
+    <Footer />
   </div>
 </template>
-
-<script>
-import Validator from './components/Validator'
-import Footer from './components/Footer'
-
-export default {
-  name: 'App',
-  components: {
-    Validator,
-    Footer
-  }
-}
-</script>
 
 <style>
 html {
