@@ -1,6 +1,8 @@
 <template>
   <footer class="footer text-center text-muted mt-5">
-    Made with &hearts; by <a href="http://fluctuo.com">fluctuo</a>, website powered by <a href="http://https://www.netlify.com/">netlify </a>
+    Made with &hearts; by <a href="http://fluctuo.com">fluctuo</a>, website
+    powered by
+    <a href="http://https://www.netlify.com/">netlify </a>
   </footer>
 </template>
 
